LayoutTests/media/video-dom-loopstart-expected.txt

Test that playing video rewinds to time specified by the loopStart DOM attribute when it loops.

EXPECTED (video.loopStart == '1') OK
EXPECTED (video.getAttribute('loopstart') == '1s') OK
RUN(video.play())
EVENT(load)
EVENT(timeupdate)
EVENT(timeupdate)
EVENT(timeupdate)
EXPECTED (video.currentLoop == '1') OK
EXPECTED (video.currentTime.toFixed(1) == '1') OK
END OF TEST
